improve useOutsideClick
