+++ frontend/common/useOutsideClick.ts
@@ -1,19 +1,25 @@
-import { RefObject, useCallback, useEffect } from 'react'
+import { RefObject, useEffect } from 'react'
 
-const useOutsideClick = (ref: RefObject<HTMLElement>, handler: () => void) => {
-  const handleClick = useCallback(
-    (event: MouseEvent) => {
-      if (ref.current && !ref.current.contains(event.target as Node)) {
-        handler()
-      }
-    },
-    [handler, ref],
-  )
-
+const useOutsideClick = (
+  ref: RefObject<HTMLElement>,
+  handler: (event: MouseEvent | TouchEvent) => void,
+) => {
   useEffect(() => {
-    document.addEventListener('click', handleClick)
-    return () => document.removeEventListener('click', handleClick)
-  }, [ref, handler, handleClick])
+    const listener = (event: MouseEvent | TouchEvent) => {
+      if (!ref.current || ref.current.contains(event.target as Node)) {
+        return
+      }
+      handler(event)
+    }
+
+    document.addEventListener('mousedown', listener)
+    document.addEventListener('touchstart', listener)
+
+    return () => {
+      document.removeEventListener('mousedown', listener)
+      document.removeEventListener('touchstart', listener)
+    }
+  }, [ref, handler])
 }
 
 export default useOutsideClick
